+++ lib/widgets/costs_list.dart
@@ -1,50 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:ugma_today/lang/localization.dart';
 import 'package:ugma_today/models/cost.dart';
-import 'package:ugma_today/routes/api_routes.dart';
-import 'package:ugma_today/utils/http/http.dart';
 
-class CostList extends StatefulWidget {
-  @override
-  _CostListState createState() => _CostListState();
-}
+class CostList extends StatelessWidget {
+  final List<Cost> costs;
 
-class _CostListState extends State<CostList> {
-  bool loading = true;
-  List<Cost> costs = [];
-  String errorMessage;
-
-  Request request;
-  @override
-  void initState() {
-    super.initState();
-
-    request = Request.get(apiRoutes.cost, context)
-      ..send()
-          .then(
-        (value) => setState(() {
-          costs = Cost.fromResponseList(value.body['data']);
-          loading = false;
-        }),
-      )
-          .catchError(
-        (error) {
-          setState(
-            () {
-              loading = false;
-              errorMessage = error;
-            },
-          );
-        },
-      );
-  }
-
-  @override
-  void dispose() {
-    request.close();
-
-    super.dispose();
-  }
+  CostList({Key key, this.costs}) : super(key: key);
 
   /// Display dialog with cost details
   void showCostDialog(BuildContext context, Cost cost) {
@@ -77,14 +38,8 @@ class _CostListState extends State<CostList> {
     );
   }
 
-  Widget showContent() {
-    if (errorMessage != null) {
-      return Text(
-        errorMessage,
-        style: Theme.of(context).textTheme.headline5,
-      );
-    }
-
+  @override
+  Widget build(BuildContext context) {
     return Container(
       constraints: BoxConstraints(maxWidth: 400.0, minWidth: 250.0),
       child: costs.length > 0
@@ -112,13 +67,6 @@ class _CostListState extends State<CostList> {
               style: Theme.of(context).textTheme.headline5,
               textAlign: TextAlign.center,
             ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: loading ? CircularProgressIndicator() : showContent(),
     );
   }
 }
